+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Il formato utilizzato è basato sulle linee guida di [Keep a Changelog](http://k
  - Aggiunta impostazione per la visualizzazione delle ore nella stampa intervento (Decimale, Sessantesimi).
  - Aggiunta possibilità di selezionare la sede di partenza della merce in fase di aggiunta articolo da un'attività
  - Sostituito plugin **Componenti** nel modulo Impianti con la possibilità di inserire gli articoli di magazzino
+ - Aggiunta colonna Scaduto nel modulo Scadenzario
  
 ### Fixed
  -